int func

+++ sem1/hyp.cpp
@@ -3,6 +3,6 @@
 #include "hyp.hpp"
 #include <cmath>
 
-double hypotenuse(int a, int b){
+int hypotenuse(int a, int b){
     return std::sqrt(a*a + b*b);
 }
